backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary
Pillow==11.3.0
python-dotenv==1.0.0
django-filter==24.3
